Line the requests up properly, and handle errors explicitly
`Plugin.Call` adds the version and plugin kind to the path, so we have
to elide those. We also have to be careful with trailing slashes,
since the plugin (if it's say, using gorilla mux) might send
redirects, and those aren't handled.

Additional: can't defer reader.Close() if it's nil, which it will be
if there's an error.